requirements.txt
beautifulsoup4
lxml
matplotlib
orjson
pymongo
PyYAML
requests
tqdm
